Fixing hooks test by adding more realistic hook module to import
